+++ src/Value/Ship/Destination.php
@@ -3,18 +3,17 @@
 namespace Phparch\SpaceTraders\Value\Ship;
 
 use Phparch\SpaceTraders\Trait\IsWaypointType;
-use Phparch\SpaceTraders\Value\SystemSymbol;
-use Phparch\SpaceTraders\Value\Waypoint\Symbol;
-use Phparch\SpaceTraders\Value\Waypoint\Type;
+use Phparch\SpaceTraders\Value\System;
+use Phparch\SpaceTraders\Value\Waypoint;
 
 class Destination
 {
     use IsWaypointType;
 
     public function __construct(
-        public readonly Symbol $symbol,
-        public readonly Type $type,
-        public readonly SystemSymbol $systemSymbol,
+        public readonly Waypoint\Symbol $symbol,
+        public readonly Waypoint\Type $type,
+        public readonly System\Symbol $systemSymbol,
         public int $x,
         public int $y,
     ) {
