+++ src/Value/Ship/Nav.php
@@ -4,16 +4,16 @@ namespace Phparch\SpaceTraders\Value\Ship;
 
 use Phparch\SpaceTraders\Trait\MapFromArray;
 use Phparch\SpaceTraders\Value\Route;
-use Phparch\SpaceTraders\Value\SystemSymbol;
-use Phparch\SpaceTraders\Value\Waypoint\Symbol;
+use Phparch\SpaceTraders\Value\System;
+use Phparch\SpaceTraders\Value\Waypoint;
 
 class Nav
 {
     use MapFromArray;
 
     public function __construct(
-        public readonly SystemSymbol $systemSymbol,
-        public readonly Symbol $waypointSymbol,
+        public readonly System\Symbol $systemSymbol,
+        public readonly Waypoint\Symbol $waypointSymbol,
         public readonly Route $route,
         public readonly Status $status,
         public readonly FlightMode $flightMode,
